moved monotonic_counter inside the implementation, and changes the base counter to the memory address of the counter, to better handle multiple monotonic_counters from different modules in the system

+++ src/inject/types.h
@@ -25,7 +25,7 @@
 
 namespace inject {
 
-typedef int unique_id;
+typedef unsigned long int unique_id;
 typedef void unknown_component;
 
 enum component_scope {
